Move PluginIntegration implementations into the proper package
