+++ plugin-compatibility/src/main/java/me/wolfyscript/utilities/compatibility/plugins/denizen/DenizenRefImpl.java
@@ -8,7 +8,6 @@ import java.io.IOException;
 import java.util.Objects;
 import me.wolfyscript.utilities.api.WolfyUtilCore;
 import me.wolfyscript.utilities.api.inventory.custom_items.references.APIReference;
-import me.wolfyscript.utilities.compatibility.plugins.DenizenIntegrationImpl;
 import me.wolfyscript.utilities.util.inventory.ItemUtils;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.Nullable;
