+++ plugin-compatibility/src/main/java/me/wolfyscript/utilities/compatibility/plugins/mmoitems/MMOItemsImpl.java
@@ -1,0 +1,52 @@
+/*
+ *       WolfyUtilities, APIs and Utilities for Minecraft Spigot plugins
+ *                      Copyright (C) 2021  WolfyScript
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package me.wolfyscript.utilities.compatibility.plugins.mmoitems;
+
+import me.wolfyscript.utilities.annotations.WUPluginIntegration;
+import me.wolfyscript.utilities.api.WolfyUtilCore;
+import me.wolfyscript.utilities.api.inventory.custom_items.references.APIReference;
+import me.wolfyscript.utilities.compatibility.PluginIntegrationAbstract;
+import org.bukkit.plugin.Plugin;
+
+@WUPluginIntegration(pluginName = MMOItemsImpl.PLUGIN_NAME)
+public class MMOItemsImpl extends PluginIntegrationAbstract {
+
+    static final String PLUGIN_NAME = "MMOItems";
+
+    protected MMOItemsImpl(WolfyUtilCore core) {
+        super(core, PLUGIN_NAME);
+    }
+
+    @Override
+    public void init(Plugin plugin) {
+        core.registerAPIReference(new MMOItemsRefImpl.Parser());
+        core.getRegistries().getStackIdentifierParsers().register(new MMOItemsStackIdentifier.Parser());
+        core.getRegistries().getStackIdentifierTypeRegistry().register(MMOItemsStackIdentifier.class);
+    }
+
+    @Override
+    public boolean hasAsyncLoading() {
+        return false;
+    }
+
+    @Override
+    public boolean isAPIReferenceIncluded(APIReference reference) {
+        return reference instanceof MMOItemsRefImpl;
+    }
+}
